+++ schedule.js
@@ -1,0 +1,12 @@
+const schedule = require('node-schedule');
+
+const etl = require('./etl');
+const log = require('./app/lib/logger');
+
+const rule = process.env.ETL_SCHEDULE || '23 * * * * *';
+
+log.info(`Scheduling job with rule '${rule}'`);
+schedule.scheduleJob(rule, () => {
+  etl.start();
+});
+
