+++ scripts/bootstrap
@@ -1,0 +1,2 @@
+#!/bin/bash
+yarn install
