+++ scripts/pre-bootstrap
@@ -1,0 +1,1 @@
+echo Please follow the instructions at https://github.com/nhsuk/dev-base-machine to set up a development environment
