+++ scripts/start
@@ -1,3 +1,3 @@
 #!/bin/bash
 docker-compose down -v
-docker-compose up --build --force-recreate full-etl
+docker-compose up --build --force-recreate full-etl nginx
