--- scripts/start-partial
@@ -1,3 +1,0 @@
-#!/bin/bash
-docker-compose down -v
-docker-compose up --build --force-recreate partial-etl
