+++ scripts/start-small
@@ -1,0 +1,3 @@
+#!/bin/bash
+docker-compose down -v
+docker-compose up --build --force-recreate partial-etl nginx
